pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice_diary"
version = "0.1.3"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "A diary that is filled in after an AI agent"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-auth>=2.38.0",
    "google-auth-oauthlib==1.2.0",
    "google-api-python-client>=2.166.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.5",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "openai>=1.14.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "pylint>=2.17.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/greet-user"
"Bug Tracker" = "https://github.com/yourusername/greet-user/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_diary"]

[tool.pytest.ini_options]
testpaths = ["src/voice_diary/db_utils/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

